--- internal/pkg/trips/reorder_item_test.go
@@ -1,35 +1,0 @@
-package trips
-
-import (
-	"github.com/DATA-DOG/go-sqlmock"
-	uuid "github.com/satori/go.uuid"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func (s *Suite) TestReorderItem_ReorderItemPosition() {
-	itemID := uuid.NewV4()
-	tripID := uuid.NewV4()
-
-	s.mock.ExpectQuery("^SELECT (.+) FROM \"items\"*").
-		WithArgs(itemID).
-		WillReturnRows(sqlmock.NewRows([]string{"id", "grocery_trip_id"}).AddRow(itemID, tripID))
-
-	s.mock.ExpectBegin()
-	s.mock.ExpectQuery("^SELECT (.+) FROM \"items\"*").
-		WithArgs(itemID).
-		WillReturnRows(sqlmock.NewRows([]string{"id"}).AddRow(itemID))
-	s.mock.ExpectExec("^UPDATE \"items\" SET (.+)$").
-		WillReturnResult(sqlmock.NewResult(1, 1))
-	s.mock.ExpectExec("^UPDATE \"grocery_trips\" SET (.+)$").
-		WillReturnResult(sqlmock.NewResult(1, 1))
-	s.mock.ExpectCommit()
-
-	s.mock.ExpectQuery("^SELECT (.+) FROM \"grocery_trips\"*").
-		WithArgs(tripID).
-		WillReturnRows(sqlmock.NewRows([]string{"id"}).AddRow(tripID))
-
-	trip, err := ReorderItem(itemID, 4)
-	require.NoError(s.T(), err)
-	assert.Equal(s.T(), trip.ID, tripID)
-}
